See comments below:
- Updates to the sos-instrument
- the PR_1 experiment was improved
- new contexts
- fixed some issues with Tika and metadata in general
- better instrumenting on the running of the predicate

+++ sos-core/src/main/java/uk/ac/standrews/cs/sos/json/MetadataDeserializer.java
@@ -50,8 +50,8 @@ public class MetadataDeserializer extends JsonDeserializer<BasicMetadata> {
     private Object getObject(JsonNode element, String type) {
 
         switch(type.toUpperCase()) {
-            case "INT":
-                return element.asInt();
+            case "LONG":
+                return element.asLong();
             case "STRING":
                 return element.asText();
             case "GUID":
